// npm/src/types/enums.ts
export enum TypeOfReturn {
	ARRAY = 'ARRAY',
	BOOLEAN = 'BOOLEAN',
	CALENDAR = 'CALENDAR',
	DURATION = 'DURATION',
	FRACTION = 'FRACTION',
	INSTANT = 'INSTANT',
	INTEGER = 'INTEGER',
	MAP = 'MAP',
	PLAIN_DATE = 'PLAIN_DATE',
	PLAIN_DATE_TIME = 'PLAIN_DATE_TIME',
	PRECISION_NUMBER = 'PRECISION_NUMBER',
	REAL_NUMBER = 'REAL_NUMBER',
	RECORD = 'RECORD',
	SET = 'SET',
	STRING = 'STRING',
	ZONED_DATE_TIME = 'ZONED_DATE_TIME',
}